+++ lib/src/util.dart
@@ -1,10 +1,8 @@
 import 'dart:math';
 import 'dart:ui';
-
 import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 import 'package:intl/intl.dart';
-
 import './models/date_header.dart';
 import './models/message_spacer.dart';
 import './models/preview_image.dart';
@@ -31,24 +29,27 @@ String getUserName(types.User user) =>
 /// Returns formatted date used as a divider between different days in the
 /// chat history
 String getVerboseDateTimeRepresentation(
-  DateTime dateTime,
-  String? locale, {
+  DateTime dateTime, {
   DateFormat? dateFormat,
+  String? dateLocale,
+  DateFormat? timeFormat,
 }) {
-  final now = DateTime.now();
+  final formattedDate = dateFormat != null
+      ? dateFormat.format(dateTime)
+      : DateFormat.MMMd(dateLocale).format(dateTime);
+  final formattedTime = timeFormat != null
+      ? timeFormat.format(dateTime)
+      : DateFormat.Hm(dateLocale).format(dateTime);
   final localDateTime = dateTime.toLocal();
-
-  if (dateFormat != null) {
-    return dateFormat.format(dateTime);
-  }
+  final now = DateTime.now();
 
   if (localDateTime.day == now.day &&
       localDateTime.month == now.month &&
       localDateTime.year == now.year) {
-    return DateFormat.Hm(locale).format(dateTime);
+    return formattedTime;
   }
 
-  return '${DateFormat.MMMd(locale).format(dateTime)}, ${DateFormat.Hm(locale).format(dateTime)}';
+  return '$formattedDate, $formattedTime';
 }
 
 /// Parses provided messages to chat messages (with headers and spacers) and
@@ -56,9 +57,11 @@ String getVerboseDateTimeRepresentation(
 List<Object> calculateChatMessages(
   List<types.Message> messages,
   types.User user, {
+  String Function(DateTime)? customDateHeaderText,
+  DateFormat? dateFormat,
   String? dateLocale,
   required bool showUserNames,
-  DateFormat? dateFormat,
+  DateFormat? timeFormat,
 }) {
   final chatMessages = <Object>[];
   final gallery = <PreviewImage>[];
@@ -120,11 +123,16 @@ List<Object> calculateChatMessages(
       chatMessages.insert(
         0,
         DateHeader(
-          text: getVerboseDateTimeRepresentation(
-            DateTime.fromMillisecondsSinceEpoch(message.createdAt!),
-            dateLocale,
-            dateFormat: dateFormat,
-          ),
+          text: customDateHeaderText != null
+              ? customDateHeaderText(
+                  DateTime.fromMillisecondsSinceEpoch(message.createdAt!),
+                )
+              : getVerboseDateTimeRepresentation(
+                  DateTime.fromMillisecondsSinceEpoch(message.createdAt!),
+                  dateFormat: dateFormat,
+                  dateLocale: dateLocale,
+                  timeFormat: timeFormat,
+                ),
         ),
       );
     }
@@ -153,11 +161,16 @@ List<Object> calculateChatMessages(
       chatMessages.insert(
         0,
         DateHeader(
-          text: getVerboseDateTimeRepresentation(
-            DateTime.fromMillisecondsSinceEpoch(nextMessage!.createdAt!),
-            dateLocale,
-            dateFormat: dateFormat,
-          ),
+          text: customDateHeaderText != null
+              ? customDateHeaderText(
+                  DateTime.fromMillisecondsSinceEpoch(nextMessage!.createdAt!),
+                )
+              : getVerboseDateTimeRepresentation(
+                  DateTime.fromMillisecondsSinceEpoch(nextMessage!.createdAt!),
+                  dateFormat: dateFormat,
+                  dateLocale: dateLocale,
+                  timeFormat: timeFormat,
+                ),
         ),
       );
     }