+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 1.1.5
 
-- Adds [onTextChanged] callback to detect when the user is typing
-- Adds [DateFormat] `customDateFormat` allowing the user to customize the date
+- Added `onTextChanged` callback to detect when user is typing. Thanks @AndreHaueisen for the PR!
+- Added `dateFormat` and `timeFormat` allowing user to customize the date and time visible between messages. Thanks @AndreHaueisen for the PR!
+- Added `customDateHeaderText` allowing to pass and arbitrary string visible between messages. See the documentation in the code for more info.
 
 ## 1.1.4